extent/v1/colf: add removed in header
But colfer couldn't support []uint16, and in theory, it has no promise of atomic load (although X86_64 has done that for single general register).

For slice, when we do atomical operation on its element, the colfer may load old data when it's doing serilization.

It's time to move header out of colfer.